+++ test/controllers/internal_controller_test.rb
@@ -1,22 +1,22 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require "test_helper"
 
 # Test to assure regular user can generate a backup.
-class InternalControllerTest < ActionController::TestCase
+class InternalControllerTest < ActionDispatch::IntegrationTest
   setup do
     @regular_user = users(:valid)
   end
 
-  test 'should get backup' do
+  test "should get backup" do
     login(@regular_user)
-    get :backup
+    get backup_url
     assert_response :success
   end
 
-  test 'should generate backup' do
+  test "should generate backup" do
     login(@regular_user)
-    post :generate_backup
-    assert_redirected_to backup_succeeded_path
+    post generate_backup_url
+    assert_redirected_to backup_succeeded_url
   end
 end