+++ test/models/charge_type_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require "test_helper"
 
 # Unit tests for charge types.
 class ChargeTypeTest < ActiveSupport::TestCase