+++ test/models/template_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require "test_helper"
 
 # Unit tests for templates.
 class TemplateTest < ActiveSupport::TestCase