+++ test/test_helper.rb
@@ -1,39 +1,34 @@
 # frozen_string_literal: true
 
-require 'simplecov'
-require 'minitest/pride'
+require "simplecov"
+require "minitest/pride"
 
-ENV['RAILS_ENV'] ||= 'test'
-require File.expand_path('../../config/environment', __FILE__)
-require 'rails/test_help'
+ENV["RAILS_ENV"] ||= "test"
+require_relative "../config/environment"
+require "rails/test_help"
 
-# Set up ActiveSupport tests
+# Set up ActiveSupport tests.
 class ActiveSupport::TestCase
-  # Setup all fixtures in test/fixtures/*.yml for all tests in alphabetical order.
+  # Setup all fixtures in test/fixtures/*.yml for all tests in alphabetical
+  # order.
   fixtures :all
 
   # Add more helper methods to be used by all tests here...
 end
 
-# Set up ActionController tests
-class ActionController::TestCase
-  include Devise::Test::ControllerHelpers
-
-  def login(resource)
-    @request.env['devise.mapping'] = Devise.mappings[resource]
-    sign_in(resource, scope: resource.class.name.downcase.to_sym)
-  end
-end
-
-# Set up ActionDispatch tests
+# Set up ActionDispatch tests.
 class ActionDispatch::IntegrationTest
+  include Devise::Test::IntegrationHelpers
+
   def login(user)
-    sign_in_as(user, '1234567890')
+    sign_in_as(user, "1234567890")
   end
 
   def sign_in_as(user, password)
     user.update password: password, password_confirmation: password
-    post '/login', params: { user: { email: user.email, password: password } }
+    post new_user_session_url, params: {
+      user: { email: user.email, password: password }
+    }
     follow_redirect!
     user
   end
@@ -41,7 +36,7 @@ end
 
 module Rack
   module Test
-    # Allow files to be uploaded in tests
+    # Allow files to be uploaded in tests.
     class UploadedFile
       attr_reader :tempfile
     end